updated maxpr.cpp
updated maxpr by sorting all the elements of the array and instead of looping just multiplied the last two elements of the sorted array.
A better O(n) implementation was to track the two max by looping through the array twice.Although my submission was successful so didn't bother to do that.